+++ packages/transaction-pay-controller/src/TransactionPayController.ts
@@ -25,8 +25,6 @@ import type {
 import { getStrategyOrder } from './utils/feature-flags';
 import { updateQuotes } from './utils/quotes';
 import { updateSourceAmounts } from './utils/source-amounts';
-import { getStrategyOrderForRoute } from './utils/strategy-routing';
-import type { TransactionPayRouteContext } from './utils/strategy-routing';
 import { pollTransactionChanges } from './utils/transaction';
 
 const MESSENGER_EXPOSED_METHODS = [
@@ -65,15 +63,10 @@ export class TransactionPayController extends BaseController<
     transaction: TransactionMeta,
   ) => TransactionPayStrategy[];
 
-  readonly #getStrategyRouteContext?: (
-    transaction: TransactionMeta,
-  ) => TransactionPayRouteContext;
-
   constructor({
     getDelegationTransaction,
     getStrategy,
     getStrategies,
-    getStrategyRouteContext,
     messenger,
     state,
   }: TransactionPayControllerOptions) {
@@ -87,7 +80,6 @@ export class TransactionPayController extends BaseController<
     this.#getDelegationTransaction = getDelegationTransaction;
     this.#getStrategy = getStrategy;
     this.#getStrategies = getStrategies;
-    this.#getStrategyRouteContext = getStrategyRouteContext;
 
     this.messenger.registerMethodActionHandlers(
       this,
@@ -282,17 +274,6 @@ export class TransactionPayController extends BaseController<
 
     if (validStrategies.length) {
       return validStrategies;
-    }
-
-    if (this.#getStrategyRouteContext) {
-      const routeStrategies = getStrategyOrderForRoute(
-        this.messenger,
-        this.#getStrategyRouteContext(transaction),
-      );
-
-      if (routeStrategies.length) {
-        return routeStrategies;
-      }
     }
 
     return getStrategyOrder(this.messenger);